+++ web/src/main/java/com/fastcms/web/config/FastcmsConfiguration.java
@@ -19,6 +19,7 @@ package com.fastcms.web.config;
 import cn.binarywang.wx.miniapp.util.WxMaConfigHolder;
 import com.fastcms.common.constants.FastcmsConstants;
 import com.fastcms.core.directive.BaseDirective;
+import com.fastcms.core.freemarker.FastcmsFreeMarkerViewResolver;
 import com.fastcms.core.site.DefaultSiteManager;
 import com.fastcms.core.site.SiteContextFilter;
 import com.fastcms.core.template.FastcmsStaticHtmlManager;
@@ -57,10 +58,7 @@ import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.i18n.AcceptHeaderLocaleResolver;
-import org.springframework.web.socket.config.annotation.EnableWebSocket;
-import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
-import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
-import org.springframework.web.socket.handler.TextWebSocketHandler;
+import org.springframework.web.servlet.view.freemarker.FreeMarkerViewResolver;
 import org.tuckey.web.filters.urlrewrite.Conf;
 import org.tuckey.web.filters.urlrewrite.UrlRewriteFilter;
 
@@ -83,8 +81,7 @@ import static com.fastcms.common.constants.FastcmsConstants.WECHAT_MINIAPP_APP_I
  * @version: 1.0
  */
 @Configuration
-@EnableWebSocket
-public class FastcmsConfiguration implements WebMvcConfigurer, WebSocketConfigurer, ApplicationListener<WebServerInitializedEvent> {
+public class FastcmsConfiguration implements WebMvcConfigurer, ApplicationListener<WebServerInitializedEvent> {
 
     @Autowired
     private IConfigService configService;
@@ -203,6 +200,12 @@ public class FastcmsConfiguration implements WebMvcConfigurer, WebSocketConfigur
         }
     }
 
+    @Bean
+    public FreeMarkerViewResolver freeMarkerViewResolver() {
+        FreeMarkerViewResolver freeMarkerViewResolver = new FastcmsFreeMarkerViewResolver();
+        return freeMarkerViewResolver;
+    }
+
     void registerFreemarkerDirective(ApplicationStartedEvent event) {
         Map<String, BaseDirective> matchingBeans = BeanFactoryUtils.beansOfTypeIncludingAncestors(event.getApplicationContext(), BaseDirective.class, true, false);
         matchingBeans.keySet().forEach(item -> {
@@ -210,11 +213,6 @@ public class FastcmsConfiguration implements WebMvcConfigurer, WebSocketConfigur
             fastcmsTemplateFreeMarkerConfig.getConfiguration().setSharedVariable(item, matchingBeans.get(item));
             pluginFreeMarkerConfig.getConfiguration().setSharedVariable(item, matchingBeans.get(item));
         });
-    }
-
-    @Override
-    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-        registry.addHandler(new TextWebSocketHandler(), "/websocket").withSockJS();
     }
 
     @Component